Suggestion chat added

+++ app/src/main/java/com/textnext/koshal/asl/MainActivity.java
@@ -1,10 +1,20 @@
 package com.textnext.koshal.asl;
 
+import android.app.ProgressDialog;
+import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.os.Vibrator;
+import android.support.design.widget.BottomSheetDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
@@ -19,6 +29,8 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.ValueEventListener;
+import com.textnext.koshal.asl.DataObjects.Feedback;
+import com.textnext.koshal.asl.DataObjects.Message;
 import com.textnext.koshal.asl.DataObjects.Online;
 import com.textnext.koshal.asl.DataObjects.Connection;
 import com.textnext.koshal.asl.DataObjects.Users;
@@ -33,8 +45,12 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
     TextView mOnline;
 
     private AdView mAdView;
+    AdRequest adRequest;
     private RewardedVideoAd mAd;
     RelativeLayout hintLayout;
+
+    ProgressDialog mProgressDialog;
+    Vibrator v;
 
 
     @Override
@@ -45,7 +61,18 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
     }
 
     private void initialize() {
+        v = (Vibrator) MainActivity.this.getSystemService(Context.VIBRATOR_SERVICE);
 
+
+        mProgressDialog = new ProgressDialog(this);
+        mProgressDialog.setOnDismissListener(new DialogInterface.OnDismissListener() {
+            @Override
+            public void onDismiss(DialogInterface dialog) {
+
+
+                ASL.mOnlineReference.child(ASL.mDeviceId).removeValue();
+            }
+        });
 
         hintLayout = (RelativeLayout) findViewById(R.id.hint_layout);
         if (ASLUtils.getShared(Constants.HINT_DISPLAYED).equals("")) {
@@ -54,7 +81,7 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
         }
 
         mAdView = (AdView) findViewById(R.id.adView);
-        AdRequest adRequest = new AdRequest.Builder().build();
+        adRequest = new AdRequest.Builder().build();
         mAdView.loadAd(adRequest);
 
         mAd = MobileAds.getRewardedVideoAdInstance(this);
@@ -68,7 +95,7 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
         mOnline = (TextView) findViewById(R.id.online);
 
 
-        if(ASL.mUser.getIsAdmin()!=null && ASL.mUser.getIsAdmin().equals("1")) {
+        if (ASL.mUser.getIsAdmin() != null && ASL.mUser.getIsAdmin().equals("1")) {
 
 
             ASL.mOnlineReference.addChildEventListener(new ChildEventListener() {
@@ -112,6 +139,9 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
         mStart.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+
+                mProgressDialog.setMessage("Searching");
+                mProgressDialog.show();
 
                 ASL.mOnlineReference.orderByChild("ts").limitToFirst(3).addListenerForSingleValueEvent(new ValueEventListener() {
 
@@ -164,27 +194,51 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
 
     private void enqueUser(int coin) {
 
-
-        ASLUtils.deleteAllMessage(MainActivity.this);
         if (coin == 0)
             ASL.mOnlineReference.child(ASL.mDeviceId).setValue(new Online(ASL.mUser.getSex(), String.valueOf(System.currentTimeMillis()), ""));
         else
             ASL.mOnlineReference.child(ASL.mDeviceId).setValue(new Online(ASL.mUser.getSex(), String.valueOf(System.currentTimeMillis()), ASL.mUser.getPriority()));
 
-        Intent i = new Intent(MainActivity.this, ChatActivity.class);
-       // i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-        startActivity(i);
-       // finish();
+
+        ASL.mDatabase.getReference().child("users").child(ASL.mDeviceId).addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                ASL.mUser = dataSnapshot.getValue(Users.class);
+
+                if (ASL.mUser.getStatus().equals("1")) {
+
+                    mProgressDialog.dismiss();
+                    v.vibrate(200);
+                    ASLUtils.setShared(Constants.CID, ASL.mUser.getCid());
+                    ASLUtils.deleteAllMessage(MainActivity.this);
+                    Intent i = new Intent(MainActivity.this, ChatActivity.class);
+                    i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                    startActivity(i);
+                    finish();
+                }
+
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+
+            }
+        });
+
+
     }
 
 
-    private void establishConnection(String uid, DataSnapshot dataSnapshot,  int coin) {
+    private void establishConnection(String uid, DataSnapshot dataSnapshot, int coin) {
+
 
         if (uid.equals(ASL.mDeviceId)) {
             enqueUser(coin);
             return;
         }
 
+        v.vibrate(200);
+        mProgressDialog.dismiss();
         dataSnapshot.getRef().removeValue();
 
         ASLUtils.setShared(Constants.P_UID, uid);
@@ -192,8 +246,9 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
         Connection connection = new Connection(uid, ASL.mDeviceId, String.valueOf(System.currentTimeMillis()), "1");
         DatabaseReference temp = ASL.mConnectionReference.push().getRef();
         String cid = temp.getKey();
-        ASLUtils.setShared(Constants.CID, cid);
         temp.setValue(connection);
+        ASLUtils.setShared(Constants.CID, cid);
+
         DatabaseReference d = ASL.mLoginReference.child(ASL.mDeviceId);
         d.child("cid").setValue(cid);
         d.child("status").setValue("1");
@@ -208,11 +263,9 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
         d2.child("status").setValue("1");
 
 
-
         ASLUtils.deleteAllMessage(MainActivity.this);
-
         Intent i = new Intent(MainActivity.this, ChatActivity.class);
-       // i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
         startActivity(i);
         finish();
     }
@@ -262,10 +315,12 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
 
     private void startPrioritySearch() {
 
+        mProgressDialog.setMessage("Searching");
+        mProgressDialog.show();
+
         ASL.mOnlineReference.orderByChild("ts").limitToLast(10).addListenerForSingleValueEvent(new ValueEventListener() {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
-
 
                 if (dataSnapshot.exists()) {
                     for (DataSnapshot d : dataSnapshot.getChildren()) {
@@ -293,4 +348,86 @@ public class MainActivity extends AppCompatActivity implements RewardedVideoAdLi
 
     }
 
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menu, menu);
+
+
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case R.id.feedback: {
+                feedbackDialog();
+                break;
+            }
+
+            case R.id.logout: {
+                logout();
+                break;
+            }
+
+        }
+        return true;
+    }
+
+
+    private void feedbackDialog() {
+        final BottomSheetDialog bottomSheetDialog = new BottomSheetDialog(MainActivity.this);
+        View view = LayoutInflater.from(this).inflate(R.layout.feedback_form_layout, null);
+        bottomSheetDialog.setContentView(view);
+        bottomSheetDialog.show();
+
+        final EditText text = (EditText) view.findViewById(R.id.feedbackText);
+        TextView send = (TextView) view.findViewById(R.id.send);
+
+        send.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                if (text.getText().toString().length() > 0) {
+
+                    Feedback feed = new Feedback();
+                    feed.setFeedback(text.getText().toString().trim());
+                    feed.setUid(ASL.mDeviceId);
+                    feed.setTs(String.valueOf(System.currentTimeMillis()));
+                    ASL.mDatabase.getReference().child("feedback").push().setValue(feed);
+                    bottomSheetDialog.dismiss();
+
+
+                } else {
+                }
+            }
+        });
+
+    }
+
+
+    private void logout() {
+
+        ASLUtils.setShared(Constants.LOGIN_STAUS, "");
+        ASLUtils.setShared(Constants.CID, "");
+        ASLUtils.setShared(Constants.P_UID, "");
+
+
+        Intent i = new Intent(MainActivity.this, LoginActivity.class);
+        i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        startActivity(i);
+        finish();
+    }
+
+    @Override
+    protected void onPause() {
+
+        if(mProgressDialog!=null)
+            mProgressDialog.dismiss();
+
+        ASL.mOnlineReference.child(ASL.mDeviceId).removeValue();
+        super.onPause();
+    }
 }
